old_project/Imageprompt_app/requirements.txt
streamlit>=1.28.0
tavily-python>=0.3.0
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
